Divided the tabs into different apps that uses signals to communicate
