+++ Apps/PlanningApp.cs
@@ -1,50 +1,10 @@
+namespace Taskly.Apps;
+using Taskly.Connections;
+using Taskly.Models;
+
 [App(icon: Icons.Calendar)]
-public class TasklyApp : ViewBase
+public class PlanningApp : ViewBase
 {
-    // Core data models
-    public record BacklogItem(
-        int Id,
-        string Title,
-        string Description,
-        int StoryPoints,
-        int Priority,
-        ItemStatus Status,
-        IssueType Type,
-        int? SprintId = null,
-        int? ParentId = null
-    );
-
-    public record Sprint(
-        int Id,
-        string Name,
-        DateTime StartDate,
-        DateTime EndDate,
-        string Goal,
-        ImmutableArray<int> ItemIds
-    );
-
-    public record Developer(
-        int Id,
-        string Name,
-        int Capacity
-    );
-
-    public enum ItemStatus
-    {
-        Backlog,
-        Todo,
-        InProgress,
-        Review,
-        Done
-    }
-
-    public enum IssueType
-    {
-        Task,
-        Bug,
-        Story,
-        Epic
-    }
 
     private static Badge GetIssueTypeBadge(IssueType type)
     {
@@ -67,6 +27,16 @@ public class TasklyApp : ViewBase
         var currentSprint = UseState(() => (Sprint)null!);
         var archivedSprints = UseState(ImmutableArray<Sprint>.Empty);
 
+        // Create signal senders for broadcasting to other apps
+        var backlogItemsSignal = Context.CreateSignal<BacklogItemsSignal, ImmutableArray<BacklogItem>, bool>();
+        var currentSprintSignal = Context.CreateSignal<CurrentSprintSignal, Sprint, bool>();
+        var archivedSprintsSignal = Context.CreateSignal<ArchivedSprintsSignal, ImmutableArray<Sprint>, bool>();
+
+        // Create signal receivers to get updates from other apps (like SprintBoardApp)
+        var backlogItemsReceiver = Context.UseSignal<BacklogItemsSignal, ImmutableArray<BacklogItem>, bool>();
+        var currentSprintReceiver = Context.UseSignal<CurrentSprintSignal, Sprint, bool>();
+        var archivedSprintsReceiver = Context.UseSignal<ArchivedSprintsSignal, ImmutableArray<Sprint>, bool>();
+
         return Layout.Vertical(
             Text.H1("Taskly - Backlog & Sprint Planning"),
             new TabsLayout(
@@ -75,14 +45,18 @@ public class TasklyApp : ViewBase
                 onRefresh: null,
                 onReorder: null,
                 selectedIndex: selectedTab.Value,
-                new Tab("Backlog", BuildBacklogView(backlogItems, currentSprint, archivedSprints)),
-                new Tab("Sprint Board", BuildSprintBoardView(backlogItems, currentSprint, archivedSprints)),
-                new Tab("Sprint Archive", BuildSprintArchiveView(archivedSprints, backlogItems, currentSprint))
+                new Tab("Backlog", BuildBacklogView(backlogItems, currentSprint, archivedSprints, backlogItemsSignal, currentSprintSignal, archivedSprintsSignal))
             )
         );
     }
 
-    private object BuildBacklogView(IState<ImmutableArray<BacklogItem>> backlogItems, IState<Sprint> currentSprint, IState<ImmutableArray<Sprint>> archivedSprints)
+    private object BuildBacklogView(
+        IState<ImmutableArray<BacklogItem>> backlogItems,
+        IState<Sprint> currentSprint,
+        IState<ImmutableArray<Sprint>> archivedSprints,
+        ISignalSender<ImmutableArray<BacklogItem>, bool> backlogItemsSignal,
+        ISignalSender<Sprint, bool> currentSprintSignal,
+        ISignalSender<ImmutableArray<Sprint>, bool> archivedSprintsSignal)
     {
         var nextId = UseState(1);
 
@@ -172,8 +146,10 @@ public class TasklyApp : ViewBase
             }
         }
 
-        void CreateSprint()
+        async ValueTask CreateSprint(Event<Button> _)
         {
+            Console.WriteLine($"CreateSprint called! Sprint name: '{newSprintName.Value}', Goal: '{newSprintGoal.Value}'");
+
             if (!string.IsNullOrWhiteSpace(newSprintName.Value))
             {
                 var sprint = new Sprint(
@@ -187,6 +163,9 @@ public class TasklyApp : ViewBase
 
                 currentSprint.Set(sprint);
                 nextSprintId.Set(nextSprintId.Value + 1);
+
+                // Broadcast the new sprint to Sprint Board app
+                await currentSprintSignal.Send(sprint);
 
                 // Clear form
                 newSprintName.Set("");
@@ -309,7 +288,7 @@ public class TasklyApp : ViewBase
             Layout.Vertical(
                 Text.H2("Product Backlog"),
 
-                new Button("+ Add Backlog Item", () => { newIssueType.Set(IssueType.Epic); isAddItemModalOpen.Set(true); }).Primary(),
+                new Button("+ Add Epic", () => { newIssueType.Set(IssueType.Epic); isAddItemModalOpen.Set(true); }).Primary(),
 
                 isAddItemModalOpen.Value ?
                     new FloatingPanel(
@@ -433,7 +412,7 @@ public class TasklyApp : ViewBase
         return Layout.Vertical(
             Text.H3($"Epics ({epics.Length})"),
             epics.Length == 0 ?
-                new Card(Text.P("No epics yet. Click '+ Add Backlog Item' to create your first epic.")) :
+                new Card(Text.P("No epics yet. Click '+ Add Epic' to create your first epic.")) :
                 Layout.Vertical(
                     epics
                         .OrderBy(x => x.Id)
@@ -736,225 +715,4 @@ public class TasklyApp : ViewBase
         );
     }
 
-
-    private object BuildSprintBoardView(IState<ImmutableArray<BacklogItem>> backlogItems, IState<Sprint> currentSprint, IState<ImmutableArray<Sprint>> archivedSprints)
-    {
-
-        // Helper method to update item status
-        void UpdateItemStatus(int id, ItemStatus newStatus)
-        {
-            var updatedItems = backlogItems.Value
-                .Select(item => item.Id == id ? item with { Status = newStatus } : item)
-                .ToImmutableArray();
-
-            backlogItems.Set(updatedItems);
-        }
-
-        void ArchiveSprint()
-        {
-            if (currentSprint.Value != null)
-            {
-                // Add current sprint to archived sprints
-                archivedSprints.Set(archivedSprints.Value.Add(currentSprint.Value));
-
-                // Clear current sprint
-                currentSprint.Set((Sprint)null!);
-            }
-        }
-
-        // Get sprint items by status
-        var sprintItems = backlogItems.Value.Where(item => currentSprint.Value != null &&
-            currentSprint.Value.ItemIds.Contains(item.Id)).ToImmutableArray();
-
-        var todoItems = sprintItems.Where(item => item.Status == ItemStatus.Todo).ToArray();
-        var inProgressItems = sprintItems.Where(item => item.Status == ItemStatus.InProgress).ToArray();
-        var doneItems = sprintItems.Where(item => item.Status == ItemStatus.Done).ToArray();
-
-        // Helper method to build sprint board item card
-        object BuildSprintBoardItemCard(BacklogItem item)
-        {
-            return new Card(
-                Layout.Vertical(
-                    // Header with issue type and title
-                    Layout.Horizontal(
-                        GetIssueTypeBadge(item.Type),
-                        Text.Strong(item.Title).Width(Size.Grow())
-                    ),
-
-                    // Description if available
-                    !string.IsNullOrEmpty(item.Description) ?
-                        Text.P(item.Description) : null,
-
-                    // Footer with story points and action buttons
-                    Layout.Horizontal(
-                        new Badge($"{item.StoryPoints} pts").Primary(),
-
-                        // Status progression buttons - forward and backward
-                        item.Status == ItemStatus.Todo ?
-                            new Button("Start", () => UpdateItemStatus(item.Id, ItemStatus.InProgress)).Primary().Small() :
-                        item.Status == ItemStatus.InProgress ?
-                            Layout.Horizontal(
-                                new Button("← Reverse", () => UpdateItemStatus(item.Id, ItemStatus.Todo)).Destructive().Small(),
-                                new Button("Complete", () => UpdateItemStatus(item.Id, ItemStatus.Done)).Primary().Small()
-                            ).Gap(4) :
-                        item.Status == ItemStatus.Done ?
-                            new Button("← Reverse", () => UpdateItemStatus(item.Id, ItemStatus.InProgress)).Destructive().Small() :
-                        null
-                    ).Gap(8)
-                ).Gap(8)
-            );
-        }
-
-        return Layout.Vertical(
-            Text.H2("Sprint Board"),
-
-            // Show current sprint info or message if no sprint
-            currentSprint.Value == null ?
-                new Card(
-                    Text.P("No active sprint. Create a sprint in the Backlog tab to get started.")
-                ) :
-                new Card(
-                    Layout.Vertical(
-                        Layout.Horizontal(
-                            Layout.Vertical(
-                                Text.H3($"Active Sprint: {currentSprint.Value.Name}"),
-                                !string.IsNullOrEmpty(currentSprint.Value.Goal) ?
-                                    Text.P($"Goal: {currentSprint.Value.Goal}") : null,
-                                Text.Small($"Total items: {sprintItems.Length} | " +
-                                          $"To Do: {todoItems.Length} | " +
-                                          $"In Progress: {inProgressItems.Length} | " +
-                                          $"Done: {doneItems.Length}")
-                            ).Width(Size.Grow()),
-                            new Button("Archive Sprint", ArchiveSprint).Secondary()
-                        )
-                    )
-                ),
-
-            // Three-column Kanban board
-            currentSprint.Value != null && sprintItems.Length > 0 ?
-                Layout.Horizontal(
-                    // To Do Column
-                    new Card(
-                        Layout.Vertical(
-                            Text.H3($"To Do ({todoItems.Length})"),
-                            todoItems.Length > 0 ?
-                                Layout.Vertical(
-                                    todoItems.Select(BuildSprintBoardItemCard).ToArray()
-                                ).Gap(8) :
-                                Text.P("No items in To Do")
-                        ).Gap(12)
-                    ).Width(Size.Grow()),
-
-                    // In Progress Column
-                    new Card(
-                        Layout.Vertical(
-                            Text.H3($"In Progress ({inProgressItems.Length})"),
-                            inProgressItems.Length > 0 ?
-                                Layout.Vertical(
-                                    inProgressItems.Select(BuildSprintBoardItemCard).ToArray()
-                                ).Gap(8) :
-                                Text.P("No items in progress")
-                        ).Gap(12)
-                    ).Width(Size.Grow()),
-
-                    // Done Column
-                    new Card(
-                        Layout.Vertical(
-                            Text.H3($"Done ({doneItems.Length})"),
-                            doneItems.Length > 0 ?
-                                Layout.Vertical(
-                                    doneItems.Select(BuildSprintBoardItemCard).ToArray()
-                                ).Gap(8) :
-                                Text.P("No completed items")
-                        ).Gap(12)
-                    ).Width(Size.Grow())
-                ).Gap(16) :
-
-                currentSprint.Value != null ?
-                    new Card(
-                        Text.P("No items in sprint yet. Add items from the Backlog tab.")
-                    ) : null
-        ).Gap(16);
-    }
-
-    private object BuildSprintArchiveView(IState<ImmutableArray<Sprint>> archivedSprints, IState<ImmutableArray<BacklogItem>> backlogItems, IState<Sprint> currentSprint)
-    {
-        void RestoreSprint(Sprint sprint)
-        {
-            // If there's a current sprint, archive it first
-            if (currentSprint.Value != null)
-            {
-                archivedSprints.Set(archivedSprints.Value.Add(currentSprint.Value));
-            }
-
-            // Remove the sprint from archives and make it current
-            archivedSprints.Set(archivedSprints.Value.Remove(sprint));
-            currentSprint.Set(sprint);
-        }
-
-        return Layout.Vertical(
-            Text.H2("Sprint Archive"),
-
-            archivedSprints.Value.Length == 0 ?
-                new Card(
-                    Text.P("No archived sprints yet. Archive a sprint from the Backlog tab to see it here.")
-                ) :
-                Layout.Vertical(
-                    archivedSprints.Value
-                        .OrderByDescending(s => s.Id)
-                        .Select(sprint =>
-                        {
-                            var sprintItems = backlogItems.Value
-                                .Where(item => sprint.ItemIds.Contains(item.Id))
-                                .ToArray();
-
-                            var completedItems = sprintItems.Where(item => item.Status == ItemStatus.Done).Count();
-                            var totalPoints = sprintItems.Sum(item => item.StoryPoints);
-                            var completedPoints = sprintItems.Where(item => item.Status == ItemStatus.Done).Sum(item => item.StoryPoints);
-
-                            return new Card(
-                                Layout.Vertical(
-                                    Layout.Horizontal(
-                                        Layout.Vertical(
-                                            Text.H3($"{sprint.Name}"),
-                                            !string.IsNullOrEmpty(sprint.Goal) ?
-                                                Text.P($"Goal: {sprint.Goal}") : null,
-                                            Text.Small($"Duration: {sprint.StartDate:MMM dd, yyyy} - {sprint.EndDate:MMM dd, yyyy}")
-                                        ).Width(Size.Grow()),
-                                        new Button("Make Current Sprint", () => RestoreSprint(sprint)).Primary()
-                                    ),
-
-                                    Layout.Horizontal(
-                                        new Badge($"{completedItems}/{sprintItems.Length} items completed").Primary(),
-                                        new Badge($"{completedPoints}/{totalPoints} points completed").Secondary()
-                                    ).Gap(4),
-
-                                    // Display sprint items
-                                    sprintItems.Length > 0 ?
-                                        Layout.Vertical(
-                                            Text.H4("Sprint Items:"),
-                                            Layout.Vertical(
-                                                sprintItems
-                                                    .OrderBy(item => item.Id)
-                                                    .Select(item => new Card(
-                                                        Layout.Horizontal(
-                                                            GetIssueTypeBadge(item.Type),
-                                                            Text.Strong(!string.IsNullOrEmpty(item.Description) ?
-                                                                $"{item.Title} - {item.Description}" : item.Title)
-                                                                .Width(Size.Grow()),
-                                                            new Badge(item.Status.ToString()).Secondary(),
-                                                            new Badge($"{item.StoryPoints} pts").Primary()
-                                                        )
-                                                    ))
-                                                    .ToArray()
-                                            ).Gap(2)
-                                        ).Gap(4) :
-                                        Text.P("No items in this sprint.")
-                                ).Gap(4)
-                            );
-                        })
-                        .ToArray()
-                ).Gap(12)
-        );
-    }
 }
